pre_retrieval/requirements.txt
sentence-transformers
pyvi
pymongo
chromadb
python-dotenv
rank-bm25
tqdm